# Fix div128 function to use Uint128 for accurate calculations in division