```
% benchstat .old.txt .new.txt                                                                                                                                                      main
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/ints
cpu: Apple M1 Pro
                         │  .old.txt   │              .new.txt              │
                         │   sec/op    │   sec/op     vs base               │
Uint256_DivMod/Uint256-8   159.6n ± 1%   143.7n ± 3%  -9.96% (p=0.000 n=10)
Uint256_DivMod/BigInt-8    75.09n ± 1%   75.27n ± 3%       ~ (p=0.239 n=10)
geomean                    109.5n        104.0n       -5.00%

                         │   .old.txt   │              .new.txt               │
                         │     B/op     │    B/op     vs base                 │
Uint256_DivMod/Uint256-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Uint256_DivMod/BigInt-8    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                               ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                         │   .old.txt   │              .new.txt               │
                         │  allocs/op   │ allocs/op   vs base                 │
Uint256_DivMod/Uint256-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Uint256_DivMod/BigInt-8    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                               ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Bug Fixes**
	- Improved accuracy of 128-bit division calculations, resulting in more reliable arithmetic operations for large numbers. No changes to user-facing interfaces.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->